initial report generation
added two new R apps, one is for file hosting, and the other is a driver for running knitr. The rook template takes YAML arguments passed by the driver to construct a pdf, and the driver saves it to the hosted file directory. The rook template is reusable apart from the driver